abstracting out useranswers in ISpecs
