+++ it/controllers/reviewAndComplete/ReviewAndCompleteControllerISpec.scala
@@ -86,47 +86,10 @@ class ReviewAndCompleteControllerISpec extends IntegrationSpecBase with CreateRe
       "user is authorised" when {
 
         "redirect to the next page" in {
-          val date = LocalDate.of(2020,1,1)
-          val deemedParentModel =
-            DeemedParentModel(
-              companyName = CompanyNameModel("Name"),
-              sautr = Some(UTRModel("1123456789")),
-              limitedLiabilityPartnership = Some(true),
-              payTaxInUk = Some(true),
-              reportingCompanySameAsParent = Some(false)
-            )
-
-          val userAnswers = UserAnswers("id")
-            .set(ReportingCompanyAppointedPage, true).get
-            .set(AgentActingOnBehalfOfCompanyPage, true).get
-            .set(AgentNamePage, "Test").get
-            .set(FullOrAbbreviatedReturnPage, Full).get
-            .set(RevisingReturnPage, true).get
-            .set(TellUsWhatHasChangedPage, "Something has changed").get
-            .set(ReportingCompanyNamePage, companyNameModel.name).get
-            .set(ReportingCompanyCTUTRPage, ctutrModel.utr).get
-            .set(AccountingPeriodPage, AccountingPeriodModel(date, date.plusMonths(1))).get
-            .set(GroupRatioElectionPage, false).get
-            .set(GroupSubjectToRestrictionsPage, true).get
-            .set(DisallowedAmountPage, BigDecimal(123.12)).get
-            .set(InterestAllowanceBroughtForwardPage, BigDecimal(123.12)).get
-            .set(GroupInterestAllowancePage, BigDecimal(123.12)).get
-            .set(GroupInterestCapacityPage, BigDecimal(123.12)).get
-            .set(EnterANGIEPage, BigDecimal(123.12)).get
-            .set(ReturnContainEstimatesPage, false).get
-            .set(ReportingCompanySameAsParentPage, false).get
-            .set(HasDeemedParentPage, false).get
-            .set(DeemedParentPage, deemedParentModel, Some(1)).get
-            .set(GroupRatioElectionPage, false).get
-            .set(ElectedInterestAllowanceAlternativeCalcBeforePage,true).get
-            .set(InterestAllowanceNonConsolidatedInvestmentsElectionPage,false).get
-            .set(ElectedInterestAllowanceConsolidatedPshipBeforePage,false).get
-            .set(InterestAllowanceConsolidatedPshipElectionPage,false).get
-
           AuthStub.authorised()
           stubPost(stubFullReturnUrl,Status.OK, Json.stringify(successfulResponse))
 
-          setAnswers(userAnswers)
+          setAnswers(fullUserAnswers)
 
           val res = postRequest("/review-and-complete", JsString(""))()
